+++ test/nimble_parsec_test.exs
@@ -14,42 +14,42 @@ defmodule NimbleParsecTest do
     @error "expected byte in the range ?0..?9, followed by byte"
 
     test "returns ok/error on composition" do
-      assert only_ascii("1a") == {:ok, [?1, ?a], "", {1, 0}, 2}
-      assert only_ascii("11") == {:ok, [?1, ?1], "", {1, 0}, 2}
-      assert only_ascii("a1") == {:error, @error, "a1", {1, 0}, 0}
+      assert only_ascii("1a") == {:ok, [?1, ?a], "", %{}, {1, 0}, 2}
+      assert only_ascii("11") == {:ok, [?1, ?1], "", %{}, {1, 0}, 2}
+      assert only_ascii("a1") == {:error, @error, "a1", %{}, {1, 0}, 0}
     end
 
     @error "expected byte in the range ?0..?9 or in the range ?z..?a"
 
     test "returns ok/error on multiple ranges" do
-      assert multi_ascii("1a") == {:ok, [?1], "a", {1, 0}, 1}
-      assert multi_ascii("a1") == {:ok, [?a], "1", {1, 0}, 1}
-      assert multi_ascii("++") == {:error, @error, "++", {1, 0}, 0}
+      assert multi_ascii("1a") == {:ok, [?1], "a", %{}, {1, 0}, 1}
+      assert multi_ascii("a1") == {:ok, [?a], "1", %{}, {1, 0}, 1}
+      assert multi_ascii("++") == {:error, @error, "++", %{}, {1, 0}, 0}
     end
 
     @error "expected byte in the range ?0..?9 or in the range ?z..?a, and not equal to ?c"
 
     test "returns ok/error on multiple ranges with not" do
-      assert multi_ascii_with_not("1a") == {:ok, [?1], "a", {1, 0}, 1}
-      assert multi_ascii_with_not("a1") == {:ok, [?a], "1", {1, 0}, 1}
-      assert multi_ascii_with_not("++") == {:error, @error, "++", {1, 0}, 0}
-      assert multi_ascii_with_not("cc") == {:error, @error, "cc", {1, 0}, 0}
+      assert multi_ascii_with_not("1a") == {:ok, [?1], "a", %{}, {1, 0}, 1}
+      assert multi_ascii_with_not("a1") == {:ok, [?a], "1", %{}, {1, 0}, 1}
+      assert multi_ascii_with_not("++") == {:error, @error, "++", %{}, {1, 0}, 0}
+      assert multi_ascii_with_not("cc") == {:error, @error, "cc", %{}, {1, 0}, 0}
     end
 
     @error "expected byte in the range ?0..?9 or in the range ?z..?a, and not equal to ?c, and not in the range ?d..?e"
 
     test "returns ok/error on multiple ranges with multiple not" do
-      assert multi_ascii_with_multi_not("1a") == {:ok, [?1], "a", {1, 0}, 1}
-      assert multi_ascii_with_multi_not("a1") == {:ok, [?a], "1", {1, 0}, 1}
-      assert multi_ascii_with_multi_not("++") == {:error, @error, "++", {1, 0}, 0}
-      assert multi_ascii_with_multi_not("cc") == {:error, @error, "cc", {1, 0}, 0}
-      assert multi_ascii_with_multi_not("de") == {:error, @error, "de", {1, 0}, 0}
+      assert multi_ascii_with_multi_not("1a") == {:ok, [?1], "a", %{}, {1, 0}, 1}
+      assert multi_ascii_with_multi_not("a1") == {:ok, [?a], "1", %{}, {1, 0}, 1}
+      assert multi_ascii_with_multi_not("++") == {:error, @error, "++", %{}, {1, 0}, 0}
+      assert multi_ascii_with_multi_not("cc") == {:error, @error, "cc", %{}, {1, 0}, 0}
+      assert multi_ascii_with_multi_not("de") == {:error, @error, "de", %{}, {1, 0}, 0}
     end
 
     test "returns ok/error even with newlines" do
-      assert ascii_newline("1a\n") == {:ok, [?1, ?a], "\n", {1, 0}, 2}
-      assert ascii_newline("1\na") == {:ok, [?1, ?\n], "a", {2, 2}, 2}
-      assert ascii_newline("\nao") == {:ok, [?\n, ?a], "o", {2, 1}, 2}
+      assert ascii_newline("1a\n") == {:ok, [?1, ?a], "\n", %{}, {1, 0}, 2}
+      assert ascii_newline("1\na") == {:ok, [?1, ?\n], "a", %{}, {2, 2}, 2}
+      assert ascii_newline("\nao") == {:ok, [?\n, ?a], "o", %{}, {2, 1}, 2}
     end
 
     test "is bound" do
@@ -65,18 +65,18 @@ defmodule NimbleParsecTest do
     @error "expected utf8 codepoint in the range ?0..?9, followed by utf8 codepoint"
 
     test "returns ok/error on composition" do
-      assert only_utf8("1a") == {:ok, [?1, ?a], "", {1, 0}, 2}
-      assert only_utf8("11") == {:ok, [?1, ?1], "", {1, 0}, 2}
-      assert only_utf8("1é") == {:ok, [?1, ?é], "", {1, 0}, 2}
-      assert only_utf8("a1") == {:error, @error, "a1", {1, 0}, 0}
+      assert only_utf8("1a") == {:ok, [?1, ?a], "", %{}, {1, 0}, 2}
+      assert only_utf8("11") == {:ok, [?1, ?1], "", %{}, {1, 0}, 2}
+      assert only_utf8("1é") == {:ok, [?1, ?é], "", %{}, {1, 0}, 2}
+      assert only_utf8("a1") == {:error, @error, "a1", %{}, {1, 0}, 0}
     end
 
     test "returns ok/error even with newlines" do
-      assert utf8_newline("1a\n") == {:ok, [?1, ?a], "\n", {1, 0}, 2}
-      assert utf8_newline("1\na") == {:ok, [?1, ?\n], "a", {2, 2}, 2}
-      assert utf8_newline("éa\n") == {:ok, [?é, ?a], "\n", {1, 0}, 2}
-      assert utf8_newline("é\na") == {:ok, [?é, ?\n], "a", {2, 2}, 2}
-      assert utf8_newline("\nao") == {:ok, [?\n, ?a], "o", {2, 1}, 2}
+      assert utf8_newline("1a\n") == {:ok, [?1, ?a], "\n", %{}, {1, 0}, 2}
+      assert utf8_newline("1\na") == {:ok, [?1, ?\n], "a", %{}, {2, 2}, 2}
+      assert utf8_newline("éa\n") == {:ok, [?é, ?a], "\n", %{}, {1, 0}, 2}
+      assert utf8_newline("é\na") == {:ok, [?é, ?\n], "a", %{}, {2, 2}, 2}
+      assert utf8_newline("\nao") == {:ok, [?\n, ?a], "o", %{}, {2, 1}, 2}
     end
 
     test "is bound" do
@@ -92,17 +92,17 @@ defmodule NimbleParsecTest do
     @error "expected byte in the range ?0..?9, followed by byte in the range ?0..?9"
 
     test "returns ok/error by itself" do
-      assert only_integer("12") == {:ok, [12], "", {1, 0}, 2}
-      assert only_integer("123") == {:ok, [12], "3", {1, 0}, 2}
-      assert only_integer("1a3") == {:error, @error, "1a3", {1, 0}, 0}
+      assert only_integer("12") == {:ok, [12], "", %{}, {1, 0}, 2}
+      assert only_integer("123") == {:ok, [12], "3", %{}, {1, 0}, 2}
+      assert only_integer("1a3") == {:error, @error, "1a3", %{}, {1, 0}, 0}
     end
 
     @error "expected string \"T\", followed by byte in the range ?0..?9, followed by byte in the range ?0..?9"
 
     test "returns ok/error with previous document" do
-      assert prefixed_integer("T12") == {:ok, ["T", 12], "", {1, 0}, 3}
-      assert prefixed_integer("T123") == {:ok, ["T", 12], "3", {1, 0}, 3}
-      assert prefixed_integer("T1a3") == {:error, @error, "T1a3", {1, 0}, 0}
+      assert prefixed_integer("T12") == {:ok, ["T", 12], "", %{}, {1, 0}, 3}
+      assert prefixed_integer("T123") == {:ok, ["T", 12], "3", %{}, {1, 0}, 3}
+      assert prefixed_integer("T1a3") == {:error, @error, "T1a3", %{}, {1, 0}, 0}
     end
 
     test "is bound" do
@@ -120,10 +120,10 @@ defmodule NimbleParsecTest do
     @error "expected byte in the range ?0..?9, followed by byte in the range ?0..?9, followed by byte in the range ?0..?9"
 
     test "returns ok/error with min" do
-      assert min_integer("123") == {:ok, [123], "", {1, 0}, 3}
-      assert min_integer("123o") == {:ok, [123], "o", {1, 0}, 3}
-      assert min_integer("1234") == {:ok, [1234], "", {1, 0}, 4}
-      assert min_integer("12") == {:error, @error, "12", {1, 0}, 0}
+      assert min_integer("123") == {:ok, [123], "", %{}, {1, 0}, 3}
+      assert min_integer("123o") == {:ok, [123], "o", %{}, {1, 0}, 3}
+      assert min_integer("1234") == {:ok, [1234], "", %{}, {1, 0}, 4}
+      assert min_integer("12") == {:error, @error, "12", %{}, {1, 0}, 0}
     end
 
     test "is not bound" do
@@ -138,17 +138,17 @@ defmodule NimbleParsecTest do
     defparsec :only_string_with_newline, string("T\nO")
 
     test "returns ok/error" do
-      assert only_string("TO") == {:ok, ["TO"], "", {1, 0}, 2}
-      assert only_string("TOC") == {:ok, ["TO"], "C", {1, 0}, 2}
-      assert only_string("AO") == {:error, "expected string \"TO\"", "AO", {1, 0}, 0}
+      assert only_string("TO") == {:ok, ["TO"], "", %{}, {1, 0}, 2}
+      assert only_string("TOC") == {:ok, ["TO"], "C", %{}, {1, 0}, 2}
+      assert only_string("AO") == {:error, "expected string \"TO\"", "AO", %{}, {1, 0}, 0}
     end
 
     test "properly counts newlines" do
-      assert only_string_with_newline("T\nO") == {:ok, ["T\nO"], "", {2, 2}, 3}
-      assert only_string_with_newline("T\nOC") == {:ok, ["T\nO"], "C", {2, 2}, 3}
+      assert only_string_with_newline("T\nO") == {:ok, ["T\nO"], "", %{}, {2, 2}, 3}
+      assert only_string_with_newline("T\nOC") == {:ok, ["T\nO"], "C", %{}, {2, 2}, 3}
 
       assert only_string_with_newline("A\nO") ==
-               {:error, "expected string \"T\\nO\"", "A\nO", {1, 0}, 0}
+               {:error, "expected string \"T\\nO\"", "A\nO", %{}, {1, 0}, 0}
     end
 
     test "is bound" do
@@ -161,17 +161,17 @@ defmodule NimbleParsecTest do
     defparsec :compile_ignore_with_newline, ignore(string("T\nO"))
 
     test "returns ok/error" do
-      assert compile_ignore("TO") == {:ok, [], "", {1, 0}, 2}
-      assert compile_ignore("TOC") == {:ok, [], "C", {1, 0}, 2}
-      assert compile_ignore("AO") == {:error, "expected string \"TO\"", "AO", {1, 0}, 0}
+      assert compile_ignore("TO") == {:ok, [], "", %{}, {1, 0}, 2}
+      assert compile_ignore("TOC") == {:ok, [], "C", %{}, {1, 0}, 2}
+      assert compile_ignore("AO") == {:error, "expected string \"TO\"", "AO", %{}, {1, 0}, 0}
     end
 
     test "properly counts newlines" do
-      assert compile_ignore_with_newline("T\nO") == {:ok, [], "", {2, 2}, 3}
-      assert compile_ignore_with_newline("T\nOC") == {:ok, [], "C", {2, 2}, 3}
+      assert compile_ignore_with_newline("T\nO") == {:ok, [], "", %{}, {2, 2}, 3}
+      assert compile_ignore_with_newline("T\nOC") == {:ok, [], "C", %{}, {2, 2}, 3}
 
       assert compile_ignore_with_newline("A\nO") ==
-               {:error, "expected string \"T\\nO\"", "A\nO", {1, 0}, 0}
+               {:error, "expected string \"T\\nO\"", "A\nO", %{}, {1, 0}, 0}
     end
 
     test "is bound" do
@@ -186,9 +186,9 @@ defmodule NimbleParsecTest do
               |> ignore()
 
     test "returns ok/error" do
-      assert runtime_ignore("abc") == {:ok, [], "", {1, 0}, 3}
+      assert runtime_ignore("abc") == {:ok, [], "", %{}, {1, 0}, 3}
       error = "expected byte in the range ?a..?z"
-      assert runtime_ignore("1bc") == {:error, error, "1bc", {1, 0}, 0}
+      assert runtime_ignore("1bc") == {:error, error, "1bc", %{}, {1, 0}, 0}
     end
 
     test "is not bound" do
@@ -202,21 +202,21 @@ defmodule NimbleParsecTest do
     defparsec :compile_replace_empty, replace(empty(), "OTHER")
 
     test "returns ok/error" do
-      assert compile_replace("TO") == {:ok, ["OTHER"], "", {1, 0}, 2}
-      assert compile_replace("TOC") == {:ok, ["OTHER"], "C", {1, 0}, 2}
-      assert compile_replace("AO") == {:error, "expected string \"TO\"", "AO", {1, 0}, 0}
+      assert compile_replace("TO") == {:ok, ["OTHER"], "", %{}, {1, 0}, 2}
+      assert compile_replace("TOC") == {:ok, ["OTHER"], "C", %{}, {1, 0}, 2}
+      assert compile_replace("AO") == {:error, "expected string \"TO\"", "AO", %{}, {1, 0}, 0}
     end
 
     test "can replace empty" do
-      assert compile_replace_empty("TO") == {:ok, ["OTHER"], "TO", {1, 0}, 0}
+      assert compile_replace_empty("TO") == {:ok, ["OTHER"], "TO", %{}, {1, 0}, 0}
     end
 
     test "properly counts newlines" do
-      assert compile_replace_with_newline("T\nO") == {:ok, ["OTHER"], "", {2, 2}, 3}
-      assert compile_replace_with_newline("T\nOC") == {:ok, ["OTHER"], "C", {2, 2}, 3}
+      assert compile_replace_with_newline("T\nO") == {:ok, ["OTHER"], "", %{}, {2, 2}, 3}
+      assert compile_replace_with_newline("T\nOC") == {:ok, ["OTHER"], "C", %{}, {2, 2}, 3}
 
       assert compile_replace_with_newline("A\nO") ==
-               {:error, "expected string \"T\\nO\"", "A\nO", {1, 0}, 0}
+               {:error, "expected string \"T\\nO\"", "A\nO", %{}, {1, 0}, 0}
     end
 
     test "is bound" do
@@ -232,9 +232,9 @@ defmodule NimbleParsecTest do
               |> replace("OTHER")
 
     test "returns ok/error" do
-      assert runtime_replace("abc") == {:ok, ["OTHER"], "", {1, 0}, 3}
+      assert runtime_replace("abc") == {:ok, ["OTHER"], "", %{}, {1, 0}, 3}
       error = "expected byte in the range ?a..?z"
-      assert runtime_replace("1bc") == {:error, error, "1bc", {1, 0}, 0}
+      assert runtime_replace("1bc") == {:error, error, "1bc", %{}, {1, 0}, 0}
     end
 
     test "is not bound" do
@@ -247,15 +247,15 @@ defmodule NimbleParsecTest do
     defparsec :compile_label_with_newline, label(string("T\nO"), "label")
 
     test "returns ok/error" do
-      assert compile_label("TO") == {:ok, ["TO"], "", {1, 0}, 2}
-      assert compile_label("TOC") == {:ok, ["TO"], "C", {1, 0}, 2}
-      assert compile_label("AO") == {:error, "expected label", "AO", {1, 0}, 0}
+      assert compile_label("TO") == {:ok, ["TO"], "", %{}, {1, 0}, 2}
+      assert compile_label("TOC") == {:ok, ["TO"], "C", %{}, {1, 0}, 2}
+      assert compile_label("AO") == {:error, "expected label", "AO", %{}, {1, 0}, 0}
     end
 
     test "properly counts newlines" do
-      assert compile_label_with_newline("T\nO") == {:ok, ["T\nO"], "", {2, 2}, 3}
-      assert compile_label_with_newline("T\nOC") == {:ok, ["T\nO"], "C", {2, 2}, 3}
-      assert compile_label_with_newline("A\nO") == {:error, "expected label", "A\nO", {1, 0}, 0}
+      assert compile_label_with_newline("T\nO") == {:ok, ["T\nO"], "", %{}, {2, 2}, 3}
+      assert compile_label_with_newline("T\nOC") == {:ok, ["T\nO"], "C", %{}, {2, 2}, 3}
+      assert compile_label_with_newline("A\nO") == {:error, "expected label", "A\nO", %{}, {1, 0}, 0}
     end
 
     test "is bound" do
@@ -271,16 +271,16 @@ defmodule NimbleParsecTest do
               |> label("third label")
 
     test "returns ok/error" do
-      assert runtime_label("aA0") == {:ok, [?a, ?A, ?0], "", {1, 0}, 3}
+      assert runtime_label("aA0") == {:ok, [?a, ?A, ?0], "", %{}, {1, 0}, 3}
 
       error = "expected first label while processing third label"
-      assert runtime_label("+A0") == {:error, error, "+A0", {1, 0}, 0}
+      assert runtime_label("+A0") == {:error, error, "+A0", %{}, {1, 0}, 0}
 
       error = "expected second label while processing third label"
-      assert runtime_label("a+0") == {:error, error, "+0", {1, 0}, 1}
+      assert runtime_label("a+0") == {:error, error, "+0", %{}, {1, 0}, 1}
 
       error = "expected third label"
-      assert runtime_label("aA+") == {:error, error, "+", {1, 0}, 2}
+      assert runtime_label("aA+") == {:error, error, "+", %{}, {1, 0}, 2}
     end
 
     test "is not bound" do
@@ -299,20 +299,20 @@ defmodule NimbleParsecTest do
 
     test "returns ok/error" do
       assert remote_runtime_traverse("T12abc34") ==
-               {:ok, ["T", 12, "99-98-97", 34], "", {1, 0}, 8}
+               {:ok, ["T", 12, "99-98-97", 34], "", %{}, {1, 0}, 8}
 
       error =
         "expected string \"T\", followed by byte in the range ?0..?9, followed by byte in the range ?0..?9"
 
-      assert remote_runtime_traverse("Tabc34") == {:error, error, "Tabc34", {1, 0}, 0}
+      assert remote_runtime_traverse("Tabc34") == {:error, error, "Tabc34", %{}, {1, 0}, 0}
 
       error = "expected byte in the range ?0..?9, followed by byte in the range ?0..?9"
-      assert remote_runtime_traverse("T12abcdf") == {:error, error, "", {1, 0}, 8}
+      assert remote_runtime_traverse("T12abcdf") == {:error, error, "", %{}, {1, 0}, 8}
 
       error =
         "expected byte in the range ?a..?z, followed by byte in the range ?a..?z, followed by byte in the range ?a..?z"
 
-      assert remote_runtime_traverse("T12ab34") == {:error, error, "ab34", {1, 0}, 3}
+      assert remote_runtime_traverse("T12ab34") == {:error, error, "ab34", %{}, {1, 0}, 3}
     end
 
     test "is not bound" do
@@ -331,20 +331,20 @@ defmodule NimbleParsecTest do
               |> integer(2)
 
     test "returns ok/error" do
-      assert local_runtime_traverse("T12abc34") == {:ok, ["T", 12, "99-98-97", 34], "", {1, 0}, 8}
+      assert local_runtime_traverse("T12abc34") == {:ok, ["T", 12, "99-98-97", 34], "", %{}, {1, 0}, 8}
 
       error =
         "expected string \"T\", followed by byte in the range ?0..?9, followed by byte in the range ?0..?9"
 
-      assert local_runtime_traverse("Tabc34") == {:error, error, "Tabc34", {1, 0}, 0}
+      assert local_runtime_traverse("Tabc34") == {:error, error, "Tabc34", %{}, {1, 0}, 0}
 
       error = "expected byte in the range ?0..?9, followed by byte in the range ?0..?9"
-      assert local_runtime_traverse("T12abcdf") == {:error, error, "", {1, 0}, 8}
+      assert local_runtime_traverse("T12abcdf") == {:error, error, "", %{}, {1, 0}, 8}
 
       error =
         "expected byte in the range ?a..?z, followed by byte in the range ?a..?z, followed by byte in the range ?a..?z"
 
-      assert local_runtime_traverse("T12ab34") == {:error, error, "ab34", {1, 0}, 3}
+      assert local_runtime_traverse("T12ab34") == {:error, error, "ab34", %{}, {1, 0}, 3}
     end
 
     test "is not bound" do
@@ -360,8 +360,8 @@ defmodule NimbleParsecTest do
               |> line()
 
     test "returns ok/error" do
-      assert ascii_line("abc") == {:ok, [{[?a, ?b, ?c], {1, 0}}], "", {1, 0}, 3}
-      assert ascii_line("a\nc") == {:ok, [{[?a, ?\n, ?c], {2, 2}}], "", {2, 2}, 3}
+      assert ascii_line("abc") == {:ok, [{[?a, ?b, ?c], {1, 0}}], "", %{}, {1, 0}, 3}
+      assert ascii_line("a\nc") == {:ok, [{[?a, ?\n, ?c], {2, 2}}], "", %{}, {2, 2}, 3}
     end
   end
 
@@ -373,8 +373,8 @@ defmodule NimbleParsecTest do
               |> byte_offset()
 
     test "returns ok/error" do
-      assert ascii_byte_offset("abc") == {:ok, [{[?a, ?b, ?c], 3}], "", {1, 0}, 3}
-      assert ascii_byte_offset("a\nc") == {:ok, [{[?a, ?\n, ?c], 3}], "", {2, 2}, 3}
+      assert ascii_byte_offset("abc") == {:ok, [{[?a, ?b, ?c], 3}], "", %{}, {1, 0}, 3}
+      assert ascii_byte_offset("a\nc") == {:ok, [{[?a, ?\n, ?c], 3}], "", %{}, {2, 2}, 3}
     end
   end
 
@@ -388,13 +388,13 @@ defmodule NimbleParsecTest do
     defparsec :empty_map, map(empty(), {Integer, :to_string, []})
 
     test "returns ok/error" do
-      assert remote_map("abc") == {:ok, ["97", "98", "99"], "", {1, 0}, 3}
-      assert remote_map("abcd") == {:ok, ["97", "98", "99"], "d", {1, 0}, 3}
-      assert {:error, _, "1abcd", {1, 0}, 0} = remote_map("1abcd")
+      assert remote_map("abc") == {:ok, ["97", "98", "99"], "", %{}, {1, 0}, 3}
+      assert remote_map("abcd") == {:ok, ["97", "98", "99"], "d", %{}, {1, 0}, 3}
+      assert {:error, _, "1abcd", %{}, {1, 0}, 0} = remote_map("1abcd")
     end
 
     test "can map empty" do
-      assert empty_map("abc") == {:ok, [], "abc", {1, 0}, 0}
+      assert empty_map("abc") == {:ok, [], "abc", %{}, {1, 0}, 0}
     end
   end
 
@@ -406,9 +406,9 @@ defmodule NimbleParsecTest do
               |> map({:local_to_string, []})
 
     test "returns ok/error" do
-      assert local_map("abc") == {:ok, ["97", "98", "99"], "", {1, 0}, 3}
-      assert local_map("abcd") == {:ok, ["97", "98", "99"], "d", {1, 0}, 3}
-      assert {:error, _, "1abcd", {1, 0}, 0} = local_map("1abcd")
+      assert local_map("abc") == {:ok, ["97", "98", "99"], "", %{}, {1, 0}, 3}
+      assert local_map("abcd") == {:ok, ["97", "98", "99"], "d", %{}, {1, 0}, 3}
+      assert {:error, _, "1abcd", %{}, {1, 0}, 0} = local_map("1abcd")
     end
 
     defp local_to_string(arg) do
@@ -426,13 +426,13 @@ defmodule NimbleParsecTest do
     defparsec :empty_reduce, reduce(empty(), {Enum, :join, ["-"]})
 
     test "returns ok/error" do
-      assert remote_reduce("abc") == {:ok, ["97-98-99"], "", {1, 0}, 3}
-      assert remote_reduce("abcd") == {:ok, ["97-98-99"], "d", {1, 0}, 3}
-      assert {:error, _, "1abcd", {1, 0}, 0} = remote_reduce("1abcd")
+      assert remote_reduce("abc") == {:ok, ["97-98-99"], "", %{}, {1, 0}, 3}
+      assert remote_reduce("abcd") == {:ok, ["97-98-99"], "d", %{}, {1, 0}, 3}
+      assert {:error, _, "1abcd", %{}, {1, 0}, 0} = remote_reduce("1abcd")
     end
 
     test "can reduce empty" do
-      assert empty_reduce("abc") == {:ok, [""], "abc", {1, 0}, 0}
+      assert empty_reduce("abc") == {:ok, [""], "abc", %{}, {1, 0}, 0}
     end
   end
 
@@ -444,9 +444,9 @@ defmodule NimbleParsecTest do
               |> reduce({:local_join, ["-"]})
 
     test "returns ok/error" do
-      assert local_reduce("abc") == {:ok, ["97-98-99"], "", {1, 0}, 3}
-      assert local_reduce("abcd") == {:ok, ["97-98-99"], "d", {1, 0}, 3}
-      assert {:error, _, "1abcd", {1, 0}, 0} = local_reduce("1abcd")
+      assert local_reduce("abc") == {:ok, ["97-98-99"], "", %{}, {1, 0}, 3}
+      assert local_reduce("abcd") == {:ok, ["97-98-99"], "d", %{}, {1, 0}, 3}
+      assert {:error, _, "1abcd", %{}, {1, 0}, 0} = local_reduce("1abcd")
     end
 
     defp local_join(list, joiner) do
@@ -462,7 +462,7 @@ defmodule NimbleParsecTest do
               )
 
     test "returns ok/error" do
-      assert concat_digit_upper_lower_plus("1Az+") == {:ok, [?1, ?A, ?z, ?+], "", {1, 0}, 4}
+      assert concat_digit_upper_lower_plus("1Az+") == {:ok, [?1, ?A, ?z, ?+], "", %{}, {1, 0}, 4}
     end
   end
 
@@ -487,24 +487,24 @@ defmodule NimbleParsecTest do
               )
 
     test "returns ok/error" do
-      assert repeat_digits("12") == {:ok, [?1, ?2], "", {1, 0}, 2}
-      assert repeat_digits("123") == {:ok, [?1, ?2], "3", {1, 0}, 2}
-      assert repeat_digits("a123") == {:ok, [], "a123", {1, 0}, 0}
+      assert repeat_digits("12") == {:ok, [?1, ?2], "", %{}, {1, 0}, 2}
+      assert repeat_digits("123") == {:ok, [?1, ?2], "3", %{}, {1, 0}, 2}
+      assert repeat_digits("a123") == {:ok, [], "a123", %{}, {1, 0}, 0}
     end
 
     test "returns ok/error with map" do
-      assert repeat_digits_to_string("123") == {:ok, ["49", "50", "51"], "", {1, 0}, 3}
+      assert repeat_digits_to_string("123") == {:ok, ["49", "50", "51"], "", %{}, {1, 0}, 3}
     end
 
     test "returns ok/error with inner and outer map" do
-      assert repeat_digits_to_same_inner("123") == {:ok, [?1, ?2, ?3], "", {1, 0}, 3}
-      assert repeat_digits_to_same_outer("123") == {:ok, [?1, ?2, ?3], "", {1, 0}, 3}
+      assert repeat_digits_to_same_inner("123") == {:ok, [?1, ?2, ?3], "", %{}, {1, 0}, 3}
+      assert repeat_digits_to_same_outer("123") == {:ok, [?1, ?2, ?3], "", %{}, {1, 0}, 3}
     end
 
     test "returns ok/error with concat map" do
-      assert repeat_double_digits_to_string("12") == {:ok, ["49", "50"], "", {1, 0}, 2}
-      assert repeat_double_digits_to_string("123") == {:ok, ["49", "50"], "3", {1, 0}, 2}
-      assert repeat_double_digits_to_string("a123") == {:ok, [], "a123", {1, 0}, 0}
+      assert repeat_double_digits_to_string("12") == {:ok, ["49", "50"], "", %{}, {1, 0}, 2}
+      assert repeat_double_digits_to_string("123") == {:ok, ["49", "50"], "3", %{}, {1, 0}, 2}
+      assert repeat_double_digits_to_string("a123") == {:ok, [], "a123", %{}, {1, 0}, 0}
     end
   end
 
@@ -531,33 +531,33 @@ defmodule NimbleParsecTest do
               )
 
     test "returns ok/error" do
-      assert repeat_while_digits("1245") == {:ok, [?1, ?2, ?4, ?5], "", {1, 0}, 4}
-      assert repeat_while_digits("12345") == {:ok, [?1, ?2], "345", {1, 0}, 2}
-      assert repeat_while_digits("135") == {:ok, [?1, ?3], "5", {1, 0}, 2}
-      assert repeat_while_digits("312") == {:ok, [], "312", {1, 0}, 0}
-      assert repeat_while_digits("a123") == {:ok, [], "a123", {1, 0}, 0}
+      assert repeat_while_digits("1245") == {:ok, [?1, ?2, ?4, ?5], "", %{}, {1, 0}, 4}
+      assert repeat_while_digits("12345") == {:ok, [?1, ?2], "345", %{}, {1, 0}, 2}
+      assert repeat_while_digits("135") == {:ok, [?1, ?3], "5", %{}, {1, 0}, 2}
+      assert repeat_while_digits("312") == {:ok, [], "312", %{}, {1, 0}, 0}
+      assert repeat_while_digits("a123") == {:ok, [], "a123", %{}, {1, 0}, 0}
     end
 
     test "returns ok/error with map" do
-      assert repeat_while_digits_to_string("123") == {:ok, ["49", "50"], "3", {1, 0}, 2}
-      assert repeat_while_digits_to_string("321") == {:ok, [], "321", {1, 0}, 0}
+      assert repeat_while_digits_to_string("123") == {:ok, ["49", "50"], "3", %{}, {1, 0}, 2}
+      assert repeat_while_digits_to_string("321") == {:ok, [], "321", %{}, {1, 0}, 0}
     end
 
     test "returns ok/error with inner and outer map" do
-      assert repeat_while_digits_to_same_inner("123") == {:ok, [?1, ?2], "3", {1, 0}, 2}
-      assert repeat_while_digits_to_same_outer("123") == {:ok, [?1, ?2], "3", {1, 0}, 2}
+      assert repeat_while_digits_to_same_inner("123") == {:ok, [?1, ?2], "3", %{}, {1, 0}, 2}
+      assert repeat_while_digits_to_same_outer("123") == {:ok, [?1, ?2], "3", %{}, {1, 0}, 2}
 
-      assert repeat_while_digits_to_same_inner("321") == {:ok, [], "321", {1, 0}, 0}
-      assert repeat_while_digits_to_same_outer("321") == {:ok, [], "321", {1, 0}, 0}
+      assert repeat_while_digits_to_same_inner("321") == {:ok, [], "321", %{}, {1, 0}, 0}
+      assert repeat_while_digits_to_same_outer("321") == {:ok, [], "321", %{}, {1, 0}, 0}
     end
 
     test "returns ok/error with concat map" do
       assert repeat_while_double_digits_to_string("12345") ==
-               {:ok, ["49", "50"], "345", {1, 0}, 2}
+               {:ok, ["49", "50"], "345", %{}, {1, 0}, 2}
 
-      assert repeat_while_double_digits_to_string("135") == {:ok, ["49", "51"], "5", {1, 0}, 2}
-      assert repeat_while_double_digits_to_string("312") == {:ok, [], "312", {1, 0}, 0}
-      assert repeat_while_double_digits_to_string("a123") == {:ok, [], "a123", {1, 0}, 0}
+      assert repeat_while_double_digits_to_string("135") == {:ok, ["49", "51"], "5", %{}, {1, 0}, 2}
+      assert repeat_while_double_digits_to_string("312") == {:ok, [], "312", %{}, {1, 0}, 0}
+      assert repeat_while_double_digits_to_string("a123") == {:ok, [], "a123", %{}, {1, 0}, 0}
     end
 
     def not_3(<<?3, _::binary>>, %{} = context, {line, line_offset}, byte_offset)
@@ -594,33 +594,33 @@ defmodule NimbleParsecTest do
               )
 
     test "returns ok/error" do
-      assert repeat_until_digits("1245") == {:ok, [?1, ?2, ?4, ?5], "", {1, 0}, 4}
-      assert repeat_until_digits("12345") == {:ok, [?1, ?2], "345", {1, 0}, 2}
-      assert repeat_until_digits("135") == {:ok, [?1, ?3], "5", {1, 0}, 2}
-      assert repeat_until_digits("312") == {:ok, [], "312", {1, 0}, 0}
-      assert repeat_until_digits("a123") == {:ok, [], "a123", {1, 0}, 0}
+      assert repeat_until_digits("1245") == {:ok, [?1, ?2, ?4, ?5], "", %{}, {1, 0}, 4}
+      assert repeat_until_digits("12345") == {:ok, [?1, ?2], "345", %{}, {1, 0}, 2}
+      assert repeat_until_digits("135") == {:ok, [?1, ?3], "5", %{}, {1, 0}, 2}
+      assert repeat_until_digits("312") == {:ok, [], "312", %{}, {1, 0}, 0}
+      assert repeat_until_digits("a123") == {:ok, [], "a123", %{}, {1, 0}, 0}
     end
 
     test "returns ok/error with map" do
-      assert repeat_until_digits_to_string("123") == {:ok, ["49", "50"], "3", {1, 0}, 2}
-      assert repeat_until_digits_to_string("321") == {:ok, [], "321", {1, 0}, 0}
+      assert repeat_until_digits_to_string("123") == {:ok, ["49", "50"], "3", %{}, {1, 0}, 2}
+      assert repeat_until_digits_to_string("321") == {:ok, [], "321", %{}, {1, 0}, 0}
     end
 
     test "returns ok/error with inner and outer map" do
-      assert repeat_until_digits_to_same_inner("123") == {:ok, [?1, ?2], "3", {1, 0}, 2}
-      assert repeat_until_digits_to_same_outer("123") == {:ok, [?1, ?2], "3", {1, 0}, 2}
+      assert repeat_until_digits_to_same_inner("123") == {:ok, [?1, ?2], "3", %{}, {1, 0}, 2}
+      assert repeat_until_digits_to_same_outer("123") == {:ok, [?1, ?2], "3", %{}, {1, 0}, 2}
 
-      assert repeat_until_digits_to_same_inner("321") == {:ok, [], "321", {1, 0}, 0}
-      assert repeat_until_digits_to_same_outer("321") == {:ok, [], "321", {1, 0}, 0}
+      assert repeat_until_digits_to_same_inner("321") == {:ok, [], "321", %{}, {1, 0}, 0}
+      assert repeat_until_digits_to_same_outer("321") == {:ok, [], "321", %{}, {1, 0}, 0}
     end
 
     test "returns ok/error with concat map" do
       assert repeat_until_double_digits_to_string("12345") ==
-               {:ok, ["49", "50"], "345", {1, 0}, 2}
+               {:ok, ["49", "50"], "345", %{}, {1, 0}, 2}
 
-      assert repeat_until_double_digits_to_string("135") == {:ok, ["49", "51"], "5", {1, 0}, 2}
-      assert repeat_until_double_digits_to_string("312") == {:ok, [], "312", {1, 0}, 0}
-      assert repeat_until_double_digits_to_string("a123") == {:ok, [], "a123", {1, 0}, 0}
+      assert repeat_until_double_digits_to_string("135") == {:ok, ["49", "51"], "5", %{}, {1, 0}, 2}
+      assert repeat_until_double_digits_to_string("312") == {:ok, [], "312", %{}, {1, 0}, 0}
+      assert repeat_until_double_digits_to_string("a123") == {:ok, [], "a123", %{}, {1, 0}, 0}
     end
   end
 
@@ -635,35 +635,35 @@ defmodule NimbleParsecTest do
               ])
 
     test "returns ok/error when bound" do
-      assert times_digits("12") == {:ok, [?1, ?2], "", {1, 0}, 2}
-      assert times_digits("123") == {:ok, [?1, ?2], "3", {1, 0}, 2}
-      assert times_digits("123456789") == {:ok, [?1, ?2, ?3, ?4, ?5, ?6, ?7, ?8], "9", {1, 0}, 8}
+      assert times_digits("12") == {:ok, [?1, ?2], "", %{}, {1, 0}, 2}
+      assert times_digits("123") == {:ok, [?1, ?2], "3", %{}, {1, 0}, 2}
+      assert times_digits("123456789") == {:ok, [?1, ?2, ?3, ?4, ?5, ?6, ?7, ?8], "9", %{}, {1, 0}, 8}
 
       assert times_digits("1234567890") ==
-               {:ok, [?1, ?2, ?3, ?4, ?5, ?6, ?7, ?8], "90", {1, 0}, 8}
+               {:ok, [?1, ?2, ?3, ?4, ?5, ?6, ?7, ?8], "90", %{}, {1, 0}, 8}
 
-      assert times_digits("12o") == {:ok, [?1, ?2], "o", {1, 0}, 2}
-      assert times_digits("o") == {:ok, [], "o", {1, 0}, 0}
+      assert times_digits("12o") == {:ok, [?1, ?2], "o", %{}, {1, 0}, 2}
+      assert times_digits("o") == {:ok, [], "o", %{}, {1, 0}, 0}
     end
 
     test "returns ok/error with choice" do
-      assert times_choice("12") == {:ok, [?1, ?2], "", {1, 0}, 2}
-      assert times_choice("123") == {:ok, [?1, ?2, ?3], "", {1, 0}, 3}
-      assert times_choice("12345") == {:ok, [?1, ?2, ?3, ?4], "5", {1, 0}, 4}
-      assert times_choice("12o") == {:ok, [?1, ?2], "o", {1, 0}, 2}
-      assert times_choice("o") == {:ok, [], "o", {1, 0}, 0}
+      assert times_choice("12") == {:ok, [?1, ?2], "", %{}, {1, 0}, 2}
+      assert times_choice("123") == {:ok, [?1, ?2, ?3], "", %{}, {1, 0}, 3}
+      assert times_choice("12345") == {:ok, [?1, ?2, ?3, ?4], "5", %{}, {1, 0}, 4}
+      assert times_choice("12o") == {:ok, [?1, ?2], "o", %{}, {1, 0}, 2}
+      assert times_choice("o") == {:ok, [], "o", %{}, {1, 0}, 0}
     end
 
     @error "expected byte in the range ?0..?9, followed by byte in the range ?0..?9 or byte in the range ?a..?z, followed by byte in the range ?a..?z"
 
     test "returns ok/error with outer choice" do
-      assert choice_times("12") == {:ok, [?1, ?2], "", {1, 0}, 2}
-      assert choice_times("12a") == {:ok, [?1, ?2], "a", {1, 0}, 2}
-      assert choice_times("12345") == {:ok, [?1, ?2, ?3, ?4], "5", {1, 0}, 4}
-      assert choice_times("ab") == {:ok, [?a, ?b], "", {1, 0}, 2}
-      assert choice_times("ab1") == {:ok, [?a, ?b], "1", {1, 0}, 2}
-      assert choice_times("abcde") == {:ok, [?a, ?b, ?c, ?d], "e", {1, 0}, 4}
-      assert choice_times("+") == {:error, @error, "+", {1, 0}, 0}
+      assert choice_times("12") == {:ok, [?1, ?2], "", %{}, {1, 0}, 2}
+      assert choice_times("12a") == {:ok, [?1, ?2], "a", %{}, {1, 0}, 2}
+      assert choice_times("12345") == {:ok, [?1, ?2, ?3, ?4], "5", %{}, {1, 0}, 4}
+      assert choice_times("ab") == {:ok, [?a, ?b], "", %{}, {1, 0}, 2}
+      assert choice_times("ab1") == {:ok, [?a, ?b], "1", %{}, {1, 0}, 2}
+      assert choice_times("abcde") == {:ok, [?a, ?b, ?c, ?d], "e", %{}, {1, 0}, 4}
+      assert choice_times("+") == {:error, @error, "+", %{}, {1, 0}, 0}
     end
   end
 
@@ -704,40 +704,40 @@ defmodule NimbleParsecTest do
     @error "expected byte in the range ?a..?z or byte in the range ?A..?Z or byte in the range ?0..?9"
 
     test "returns ok/error" do
-      assert simple_choice("a=") == {:ok, [?a], "=", {1, 0}, 1}
-      assert simple_choice("A=") == {:ok, [?A], "=", {1, 0}, 1}
-      assert simple_choice("0=") == {:ok, [?0], "=", {1, 0}, 1}
-      assert simple_choice("+=") == {:error, @error, "+=", {1, 0}, 0}
+      assert simple_choice("a=") == {:ok, [?a], "=", %{}, {1, 0}, 1}
+      assert simple_choice("A=") == {:ok, [?A], "=", %{}, {1, 0}, 1}
+      assert simple_choice("0=") == {:ok, [?0], "=", %{}, {1, 0}, 1}
+      assert simple_choice("+=") == {:error, @error, "+=", %{}, {1, 0}, 0}
     end
 
     test "returns ok/error with repeat inside" do
-      assert choice_inner_repeat("az") == {:ok, [?a, ?z], "", {1, 0}, 2}
-      assert choice_inner_repeat("AZ") == {:ok, [], "AZ", {1, 0}, 0}
+      assert choice_inner_repeat("az") == {:ok, [?a, ?z], "", %{}, {1, 0}, 2}
+      assert choice_inner_repeat("AZ") == {:ok, [], "AZ", %{}, {1, 0}, 0}
     end
 
     test "returns ok/error with repeat outside" do
-      assert choice_outer_repeat("az") == {:ok, [?a, ?z], "", {1, 0}, 2}
-      assert choice_outer_repeat("AZ") == {:ok, [?A, ?Z], "", {1, 0}, 2}
-      assert choice_outer_repeat("aAzZ") == {:ok, [?a, ?A, ?z, ?Z], "", {1, 0}, 4}
+      assert choice_outer_repeat("az") == {:ok, [?a, ?z], "", %{}, {1, 0}, 2}
+      assert choice_outer_repeat("AZ") == {:ok, [?A, ?Z], "", %{}, {1, 0}, 2}
+      assert choice_outer_repeat("aAzZ") == {:ok, [?a, ?A, ?z, ?Z], "", %{}, {1, 0}, 4}
     end
 
     test "returns ok/error with repeat and inner map" do
-      assert choice_repeat_and_inner_map("az") == {:ok, ["97", "122"], "", {1, 0}, 2}
-      assert choice_repeat_and_inner_map("AZ") == {:ok, ["65", "90"], "", {1, 0}, 2}
+      assert choice_repeat_and_inner_map("az") == {:ok, ["97", "122"], "", %{}, {1, 0}, 2}
+      assert choice_repeat_and_inner_map("AZ") == {:ok, ["65", "90"], "", %{}, {1, 0}, 2}
 
       assert choice_repeat_and_inner_map("aAzZ") ==
-               {:ok, ["97", "65", "122", "90"], "", {1, 0}, 4}
+               {:ok, ["97", "65", "122", "90"], "", %{}, {1, 0}, 4}
     end
 
     test "returns ok/error with repeat and maps" do
-      assert choice_repeat_and_maps("az") == {:ok, [?a, ?z], "", {1, 0}, 2}
-      assert choice_repeat_and_maps("AZ") == {:ok, [?A, ?Z], "", {1, 0}, 2}
-      assert choice_repeat_and_maps("aAzZ") == {:ok, [?a, ?A, ?z, ?Z], "", {1, 0}, 4}
+      assert choice_repeat_and_maps("az") == {:ok, [?a, ?z], "", %{}, {1, 0}, 2}
+      assert choice_repeat_and_maps("AZ") == {:ok, [?A, ?Z], "", %{}, {1, 0}, 2}
+      assert choice_repeat_and_maps("aAzZ") == {:ok, [?a, ?A, ?z, ?Z], "", %{}, {1, 0}, 4}
     end
 
     test "returns ok/error on empty" do
-      assert choice_with_empty("az") == {:ok, [?a], "z", {1, 0}, 1}
-      assert choice_with_empty("AZ") == {:ok, [], "AZ", {1, 0}, 0}
+      assert choice_with_empty("az") == {:ok, [?a], "z", %{}, {1, 0}, 1}
+      assert choice_with_empty("AZ") == {:ok, [], "AZ", %{}, {1, 0}, 0}
     end
   end
 
@@ -745,8 +745,8 @@ defmodule NimbleParsecTest do
     defparsec :optional_ascii, optional(ascii_char([?a..?z]))
 
     test "returns ok/error on empty" do
-      assert optional_ascii("az") == {:ok, [?a], "z", {1, 0}, 1}
-      assert optional_ascii("AZ") == {:ok, [], "AZ", {1, 0}, 0}
+      assert optional_ascii("az") == {:ok, [?a], "z", %{}, {1, 0}, 1}
+      assert optional_ascii("AZ") == {:ok, [], "AZ", %{}, {1, 0}, 0}
     end
   end
 
@@ -763,39 +763,39 @@ defmodule NimbleParsecTest do
     defparsec :parsec_choice, choice([parsec(:parsec_inner), string("+")])
 
     test "returns ok/error with string" do
-      assert parsec_string("TaO") == {:ok, ["T", "97", "O"], "", {1, 0}, 3}
+      assert parsec_string("TaO") == {:ok, ["T", "97", "O"], "", %{}, {1, 0}, 3}
 
       error = "expected string \"T\""
-      assert parsec_string("ZaO") == {:error, error, "ZaO", {1, 0}, 0}
+      assert parsec_string("ZaO") == {:error, error, "ZaO", %{}, {1, 0}, 0}
 
       error = "expected byte in the range ?a..?z or byte in the range ?A..?Z"
-      assert parsec_string("T1O") == {:error, error, "1O", {1, 0}, 1}
+      assert parsec_string("T1O") == {:error, error, "1O", %{}, {1, 0}, 1}
 
       error = "expected string \"O\""
-      assert parsec_string("TaA") == {:error, error, "A", {1, 0}, 2}
+      assert parsec_string("TaA") == {:error, error, "A", %{}, {1, 0}, 2}
     end
 
     test "returns ok/error with choice" do
-      assert parsec_choice("+O") == {:ok, ["+"], "O", {1, 0}, 1}
-      assert parsec_choice("O+") == {:ok, ["79"], "+", {1, 0}, 1}
+      assert parsec_choice("+O") == {:ok, ["+"], "O", %{}, {1, 0}, 1}
+      assert parsec_choice("O+") == {:ok, ["79"], "+", %{}, {1, 0}, 1}
 
       assert parsec_choice("==") ==
-               {:error, "expected parsec_inner or string \"+\"", "==", {1, 0}, 0}
+               {:error, "expected parsec_inner or string \"+\"", "==", %{}, {1, 0}, 0}
     end
 
     test "returns ok/error with repeat" do
-      assert parsec_repeat("az") == {:ok, ["97", "122"], "", {1, 0}, 2}
-      assert parsec_repeat("AZ") == {:ok, ["65", "90"], "", {1, 0}, 2}
-      assert parsec_repeat("aAzZ") == {:ok, ["97", "65", "122", "90"], "", {1, 0}, 4}
-      assert parsec_repeat("1aAzZ") == {:ok, [], "1aAzZ", {1, 0}, 0}
+      assert parsec_repeat("az") == {:ok, ["97", "122"], "", %{}, {1, 0}, 2}
+      assert parsec_repeat("AZ") == {:ok, ["65", "90"], "", %{}, {1, 0}, 2}
+      assert parsec_repeat("aAzZ") == {:ok, ["97", "65", "122", "90"], "", %{}, {1, 0}, 4}
+      assert parsec_repeat("1aAzZ") == {:ok, [], "1aAzZ", %{}, {1, 0}, 0}
     end
 
     @error "expected byte in the range ?a..?z or byte in the range ?A..?Z"
 
     test "returns ok/error with map" do
-      assert parsec_map("az") == {:ok, [?a], "z", {1, 0}, 1}
-      assert parsec_map("AZ") == {:ok, [?A], "Z", {1, 0}, 1}
-      assert parsec_map("1aAzZ") == {:error, @error, "1aAzZ", {1, 0}, 0}
+      assert parsec_map("az") == {:ok, [?a], "z", %{}, {1, 0}, 1}
+      assert parsec_map("AZ") == {:ok, [?A], "Z", %{}, {1, 0}, 1}
+      assert parsec_map("1aAzZ") == {:error, @error, "1aAzZ", %{}, {1, 0}, 0}
     end
   end
 
@@ -817,7 +817,7 @@ defmodule NimbleParsecTest do
     defparsec :datetime, date |> ignore(string("T")) |> concat(time), inline: true
 
     test "returns ok/error by itself" do
-      assert datetime("2010-04-17T14:12:34") == {:ok, [2010, 4, 17, 14, 12, 34], "", {1, 0}, 19}
+      assert datetime("2010-04-17T14:12:34") == {:ok, [2010, 4, 17, 14, 12, 34], "", %{}, {1, 0}, 19}
     end
   end